# docs/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Bardic server API
  version: 0.2.6
  summary: Contract between the Bardic server and its clients (draft).
  description: |
    Normative, language-neutral contract for Bardic v2. It is derived from
    `docs/PRODUCT-SPEC.md`, not from the prototype. Clients are generated
    from this file; the server must satisfy it. Read `README.md` in this folder
    first: it explains identity headers, errors, money, events, long-running
    work and the place-sync protocol once, so each operation does not repeat them.

    Status: pre-release (0.x). Until 1.0.0, additive changes bump the patch
    version and breaking changes bump the minor version.
  license:
    name: Personal use
    identifier: LicenseRef-Personal
servers:
  - url: http://localhost:8765
    description: Default local address. The server name and address are configurable.
...
paths:
...
  /api/audiobooks/{audiobook_id}/space:
    parameters: [{$ref: '#/components/parameters/AudiobookId'}]
    get:
      operationId: getAudiobookSpace
      tags: [Audiobooks]
      summary: Space used by an audiobook and what re-making would cost
      description: Drives the free-up-space sheet. Read-only.
      x-bardic-cost: network
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Usage.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AudiobookSpace'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      operationId: freeAudiobookSpace
      tags: [Audiobooks]
      summary: Delete an audiobook's regenerable audio
      description: |
        Deletes audio on the server; keeps the book, places, plans, audit and
        downloads on devices. Returns the freed bytes. Refuses while a job is
        running for it (409 `job_running`). The records of the freed audio (size,
        revision, duration) are kept, so `checkDownloads` can still describe a
        copy a device holds. For a premium voice, see `getAudiobookSpace` for what
        making it again costs.
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Result.
          content:
            application/json:
              schema:
                type: object
                required: [freed_bytes]
                properties:
                  freed_bytes: {type: integer, minimum: 0}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/audiobooks/{audiobook_id}/manifest:
    parameters: [{$ref: '#/components/parameters/AudiobookId'}]
    get:
      operationId: getAudiobookManifest
      tags: [Offline]
      summary: Everything a device needs to download an audiobook
      description: |
        For every ready chapter: audio id, size, hash, duration, the timings id
        and the text hash. Clients download `getAudio`, `getAudioTimings` and
        `getChapterText` for the chapters they choose.
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Manifest.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Manifest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/audiobooks/{audiobook_id}/sync-check:
    parameters: [{$ref: '#/components/parameters/AudiobookId'}]
    post:
      operationId: checkDownloads
      tags: [Offline]
      summary: Which downloaded chapters are out of date
      description: |
        The device sends what it holds. For every chapter whose audio has been
        made again (later, by the same voice, in this or a newer audiobook of the
        book), the response says what changed, so the client can show the
        comparison and let the listener decide. Read-only; nothing is replaced.
        A held copy whose server copy was only freed, or that this server has
        never heard of, is listed as up to date: a kept copy keeps playing.
        404 `chapter_not_found` for a chapter of another book.
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [have]
              properties:
                have:
                  type: array
                  items:
                    type: object
                    required: [chapter_id, audio_id]
                    properties:
                      chapter_id: {type: string}
                      audio_id: {type: string}
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Result.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SyncCheck'}
        '404': {$ref: '#/components/responses/NotFound'}

  # ------------------------------------------------------------ Admin
...
components:
  parameters:
...
    Device:
      name: X-Bardic-Device
      in: header
      required: true
      description: Client-generated device id (UUID). Registers the device on first use and is recorded in the audit log.
      schema: {type: string}
...
    AudiobookId: {name: audiobook_id, in: path, required: true, schema: {type: string}}
...
  responses:
    InvalidRequest:
      description: The request is malformed or breaks a rule; `code` says which.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: The resource does not exist (or is not visible to this listener).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: The request conflicts with current state; `code` says how.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    ServerError:
      description: An unexpected failure on the server. Nothing you need to undo was changed; `retryable` is true.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
...
    Error:
      type: object
      required: [code, detail]
      properties:
        code: {type: string, description: "Stable machine-readable code (open enumeration); see README."}
        detail: {type: string, description: "Human-readable, safe to show."}
        retryable: {type: boolean}
        retry_after_seconds: {type: [integer, 'null']}
        context:
          type: object
          additionalProperties: true
          description: Code-specific fields (for example `plan_id`, `existing_book_id`).
    Money:
      type: object
      required: [micros, currency]
      properties:
        micros: {type: integer, description: "Millionths of the currency unit. Never a float."}
        currency: {type: string, pattern: '^[A-Z]{3}$', examples: [USD]}
...
    AudioRef:
      type: object
      required: [id, duration_seconds, bytes, sha256, content_type, voice_revision]
      properties:
        id: {type: string}
        duration_seconds: {type: number}
        bytes: {type: integer}
        sha256: {type: string}
        content_type: {type: string}
        voice_revision: {type: string}
        url: {type: string}
        timings_url: {type: string}
...
    AudiobookSpace:
      type: object
      required: [audiobook_id, bytes, chapters, remake_estimate]
      properties:
        audiobook_id: {type: string}
        bytes: {type: integer}
        chapters: {type: integer}
        remake_estimate:
          description: What making it again would cost; null for free voices.
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/CostEstimate'
...
    CostEstimate:
      type: object
      required: [low, likely, high, prices_as_of, basis]
      properties:
        low: {$ref: '#/components/schemas/Money'}
        likely: {$ref: '#/components/schemas/Money'}
        high: {$ref: '#/components/schemas/Money'}
        prices_as_of: {type: string, format: date-time}
        basis: {type: string, enum: [provider, manual, unknown], description: "Where the price came from; `unknown` means no range can be given."}
...
    Manifest:
      type: object
      required: [audiobook_id, generated_at, chapters]
      properties:
        audiobook_id: {type: string}
        generated_at: {type: string, format: date-time}
        chapters:
          type: array
          items:
            type: object
            required: [chapter_id, audio, text_sha256]
            properties:
              chapter_id: {type: string}
              audio: {$ref: '#/components/schemas/AudioRef'}
              text_sha256: {type: string}
    SyncCheck:
      type: object
      required: [up_to_date, newer]
      properties:
        up_to_date:
          type: array
          items: {type: string}
          description: Chapter ids whose device copy is current.
        newer:
          type: array
          items:
            type: object
            required: [chapter_id, old_audio_id, new_audio, changes]
            properties:
              chapter_id: {type: string}
              old_audio_id: {type: string}
              new_audio: {$ref: '#/components/schemas/AudioRef'}
              changes:
                type: object
                required: [voice_name, old_voice_revision, new_voice_revision, old_duration_seconds, new_duration_seconds, old_bytes, new_bytes]
                properties:
                  voice_name: {type: string}
                  old_voice_revision: {type: string}
                  new_voice_revision: {type: string}
                  old_duration_seconds: {type: number}
                  new_duration_seconds: {type: number}
                  old_bytes: {type: integer}
                  new_bytes: {type: integer}

    # ---- admin
...